Allow user to override pages with arbitrary IDs
From the widget screen, a user can optionally specify a list of page IDs
(or `$post->post_name` or in the case of an archive, the string
`archive:xxxxxx` where `xxxxx` is a post type) to override the default
behavior.

This commit introduces some procedural code disguised as classes. Future
commits should be aim to provide better abstractions.